refactor: update database schema and improve CanteenItem model
- Refactored SQL schema for various tables to ensure consistency and clarity.
- Changed product_price type in CanteenItem model from FLOAT to DECIMAL(10, 2) for better precision.
- Updated CanteenItemCard component for improved formatting and user interaction.
- Adjusted router configuration to point to the correct Canteen component.

+++ back-node/node/models/CanteenItem.js
@@ -4,7 +4,7 @@ import sequelize from '../config/database.js';
 const CanteenItem = sequelize.define('CanteenItem', {
     id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true, },
     product_name: { type: DataTypes.TEXT, allowNull: false, },
-    product_price: { type: DataTypes.FLOAT, allowNull: false, },
+    product_price: { type: DataTypes.DECIMAL(10, 2), allowNull: false, },
     product_enabled: { type: DataTypes.BOOLEAN, allowNull: false, defaultValue: true },
     createdAt: { field: 'created_at', type: DataTypes.DATE },
     updatedAt: { field: 'updated_at', type: DataTypes.DATE },
